move function out + add documentation
